backend/pyproject.toml
[project]
name = "nexus-financial-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "langchain[openai]>=1.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
